requirements.txt
alembic==1.1.0
blinker==1.4
Click==7.0
configparser==3.8.1
Flask==1.1.1
Flask-JSON==0.3.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.1.2
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
SQLAlchemy-serializer==1.3.4.2
Werkzeug==0.15.5
WTForms==2.2.1
